# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canlı Maç Takip & Analiz API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Team {
  id: number;
  name: string;
  shortName: string;
  logo: string;
  country: string;
}

export interface League {
  id: number;
  name: string;
  country: string;
  logo: string;
  season: string;
}

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  live: "live",
  upcoming: "upcoming",
  finished: "finished",
  halftime: "halftime",
} as const;

export interface Match {
  id: number;
  leagueId: number;
  leagueName: string;
  leagueLogo: string;
  homeTeam: Team;
  awayTeam: Team;
  /** @nullable */
  homeScore: number | null;
  /** @nullable */
  awayScore: number | null;
  status: MatchStatus;
  /** @nullable */
  minute: number | null;
  startTime: string;
  date: string;
}

export interface MatchStats {
  homePossession: number;
  awayPossession: number;
  homeShots: number;
  awayShots: number;
  homeShotsOnTarget: number;
  awayShotsOnTarget: number;
  homeCorners: number;
  awayCorners: number;
  homeFouls: number;
  awayFouls: number;
  homeYellowCards: number;
  awayYellowCards: number;
  homeRedCards: number;
  awayRedCards: number;
  homeOffsides: number;
  awayOffsides: number;
}

export type MatchDetailStatus =
  (typeof MatchDetailStatus)[keyof typeof MatchDetailStatus];

export const MatchDetailStatus = {
  live: "live",
  upcoming: "upcoming",
  finished: "finished",
  halftime: "halftime",
} as const;

export interface MatchDetail {
  id: number;
  leagueId: number;
  leagueName: string;
  leagueLogo: string;
  homeTeam: Team;
  awayTeam: Team;
  /** @nullable */
  homeScore: number | null;
  /** @nullable */
  awayScore: number | null;
  status: MatchDetailStatus;
  /** @nullable */
  minute: number | null;
  startTime: string;
  date: string;
  stats: MatchStats;
}

export type MatchEventType =
  (typeof MatchEventType)[keyof typeof MatchEventType];

export const MatchEventType = {
  goal: "goal",
  yellow_card: "yellow_card",
  red_card: "red_card",
  substitution: "substitution",
  penalty: "penalty",
  own_goal: "own_goal",
} as const;

export type MatchEventTeam =
  (typeof MatchEventTeam)[keyof typeof MatchEventTeam];

export const MatchEventTeam = {
  home: "home",
  away: "away",
} as const;

export interface MatchEvent {
  id: number;
  matchId: number;
  minute: number;
  type: MatchEventType;
  team: MatchEventTeam;
  playerName: string;
  /** @nullable */
  assistName: string | null;
  /** @nullable */
  description: string | null;
}

export interface MatchSummary {
  liveCount: number;
  upcomingCount: number;
  finishedCount: number;
  totalGoalsToday: number;
}

export interface Standing {
  rank: number;
  team: Team;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  goalDifference: number;
  points: number;
  form: string;
}

export type PredictionConfidence =
  (typeof PredictionConfidence)[keyof typeof PredictionConfidence];

export const PredictionConfidence = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface Prediction {
  matchId: number;
  homeWinProbability: number;
  drawProbability: number;
  awayWinProbability: number;
  predictedScore: string;
  confidence: PredictionConfidence;
  analysis: string;
  btts: boolean;
  over25: boolean;
}

export type ListMatchesParams = {
  status?: ListMatchesStatus;
  /**
   * @nullable
   */
  leagueId?: number | null;
};

export type ListMatchesStatus =
  (typeof ListMatchesStatus)[keyof typeof ListMatchesStatus];

export const ListMatchesStatus = {
  live: "live",
  upcoming: "upcoming",
  finished: "finished",
  all: "all",
} as const;