# lib/api-zod/src/generated/types/league.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canlı Maç Takip & Analiz API
 * OpenAPI spec version: 0.1.0
 */

export interface League {
  id: number;
  name: string;
  country: string;
  logo: string;
  season: string;
}