# lib/api-zod/src/generated/types/matchDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canlı Maç Takip & Analiz API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchDetailStatus } from "./matchDetailStatus";
import type { MatchStats } from "./matchStats";
import type { Team } from "./team";

export interface MatchDetail {
  id: number;
  leagueId: number;
  leagueName: string;
  leagueLogo: string;
  homeTeam: Team;
  awayTeam: Team;
  /** @nullable */
  homeScore: number | null;
  /** @nullable */
  awayScore: number | null;
  status: MatchDetailStatus;
  /** @nullable */
  minute: number | null;
  startTime: string;
  date: string;
  stats: MatchStats;
}