# lib/api-zod/src/generated/types/matchEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canlı Maç Takip & Analiz API
 * OpenAPI spec version: 0.1.0
 */
import type { MatchEventTeam } from "./matchEventTeam";
import type { MatchEventType } from "./matchEventType";

export interface MatchEvent {
  id: number;
  matchId: number;
  minute: number;
  type: MatchEventType;
  team: MatchEventTeam;
  playerName: string;
  /** @nullable */
  assistName: string | null;
  /** @nullable */
  description: string | null;
}