# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canlı Maç Takip & Analiz API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionConfidence } from "./predictionConfidence";

export interface Prediction {
  matchId: number;
  homeWinProbability: number;
  drawProbability: number;
  awayWinProbability: number;
  predictedScore: string;
  confidence: PredictionConfidence;
  analysis: string;
  btts: boolean;
  over25: boolean;
}