# lib/api-zod/src/generated/types/standing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canlı Maç Takip & Analiz API
 * OpenAPI spec version: 0.1.0
 */
import type { Team } from "./team";

export interface Standing {
  rank: number;
  team: Team;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  goalDifference: number;
  points: number;
  form: string;
}