# lib/api-zod/src/generated/types/team.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Canlı Maç Takip & Analiz API
 * OpenAPI spec version: 0.1.0
 */

export interface Team {
  id: number;
  name: string;
  shortName: string;
  logo: string;
  country: string;
}